Initial commit that notifies admins about new users
This bot receives team_join Slack events and sends a message with
the new user's username and email to the private #slack-admins
channel. It requires the application that it is talking to (which
is specified by the SLACK_BOT_TOKEN in a config.json file) to have
a bot user (with the bot user's OAuth token in said config.json)
that has been added to the #slack-admins channel. It should also
be added to the #bots channel for debugging. :)